test: extract shared subgraph fixture helpers to fix sonar duplication
