+++ test/subgraph-headers.e2e-spec.ts
@@ -1,76 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
-import * as http from 'node:http';
-import { AddressInfo } from 'node:net';
 import { AppModule } from './../src/app.module';
-
-function makeSdl(fieldName: string): string {
-  return `
-extend schema
-  @link(url: "https://specs.apollo.dev/federation/v2.3", import: ["@key"])
-
-type Query {
-  ${fieldName}: String
-}
-`;
-}
-
-type IncomingRequest = {
-  url?: string;
-  headers: http.IncomingHttpHeaders;
-  body: string;
-};
-
-function recordRequest(
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-  requests: IncomingRequest[],
-  sdl: string,
-): void {
-  const chunks: Buffer[] = [];
-  req.on('data', (chunk: Buffer) => chunks.push(chunk));
-  req.on('end', () => {
-    requests.push({
-      url: req.url,
-      headers: req.headers,
-      body: Buffer.concat(chunks).toString('utf-8'),
-    });
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.end(JSON.stringify({ data: { _service: { sdl } } }));
-  });
-}
-
-function closeServer(server: http.Server): Promise<void> {
-  return new Promise<void>((done) => server.close(() => done()));
-}
-
-function listenOnRandomPort(server: http.Server): Promise<number> {
-  return new Promise((resolve) => {
-    server.listen(0, '127.0.0.1', () => {
-      const { port } = server.address() as AddressInfo;
-      resolve(port);
-    });
-  });
-}
-
-async function startSubgraphFixture(fieldName: string = 'hello'): Promise<{
-  port: number;
-  requests: IncomingRequest[];
-  close: () => Promise<void>;
-}> {
-  const requests: IncomingRequest[] = [];
-  const sdl = makeSdl(fieldName);
-  const server = http.createServer((req, res) =>
-    recordRequest(req, res, requests, sdl),
-  );
-  const port = await listenOnRandomPort(server);
-  return { port, requests, close: () => closeServer(server) };
-}
+import {
+  startSubgraphFixture,
+  SubgraphFixture,
+} from './helpers/subgraph-fixture';
 
 describe('Subgraph headers (e2e)', () => {
   const ORIGINAL_ENV = process.env;
-  let fixture: Awaited<ReturnType<typeof startSubgraphFixture>> | undefined;
+  let fixture: SubgraphFixture | undefined;
   let app: INestApplication | undefined;
 
   beforeEach(async () => {
